Make skydoc compatible with Python 3.
Also removes the dependency on python-gflags by replacing it with
optparse, which is available in the standard library of Python 2 and 3.

Where possible I wrote the code so that it doesn't even require Python
2.7, but should still work with even older versions. (OrderedDict only
ships starting with Python 2.7).

This also fixes a bug in skydoc in which the ordering of documentation
for outputs wasn't deterministic (because dicts in Python 2 didn't
preserve insertion order by default).

Tested with:

bazel test //...

bazel test --python_path=/Users/philwo/homebrew/bin/python3.7 //...